api/requirements.txt
appdirs==1.4.3
astroid==1.5.3
Babel==2.5.1
cached-property==1.3.1
certifi==2017.11.5
chardet==3.0.4
click==6.7
colorama==0.3.9
coreapi==2.3.3
coreschema==0.0.4
debtcollector==1.18.0
Django==1.11.7
django-debug-toolbar==1.9
django-extensions==1.9.7
django-filter==1.1.0
django-filters==0.2.1
django-rest-swagger==2.1.2
djangorestframework==3.7.3
djangorestframework-gis==0.12
docker==2.6.1
docker-compose==1.17.1
docker-pycreds==0.2.1
dockerpty==0.4.1
docopt==0.6.2
drf-extensions==0.3.1
elasticsearch==5.5.1
elasticsearch-dsl==5.3.0
factory-boy==2.9.2
Faker==0.8.6
first==2.0.1
flake8==3.5.0
gevent==1.2.2
graypy==0.2.14
greenlet==0.4.12
grequests==0.3.0
idna==2.6
iso8601==0.1.12
isort==4.2.15
itypes==1.1.0
Jinja2==2.10
jsonschema==2.6.0
keystoneauth1==3.2.0
lazy-object-proxy==1.3.1
Markdown==2.6.9
MarkupSafe==1.0
mccabe==0.6.1
monotonic==1.4
msgpack-python==0.4.8
netaddr==0.7.19
netifaces==0.10.6
openapi-codec==1.3.2
oslo.config==5.0.0
oslo.i18n==3.18.0
oslo.serialization==2.21.2
oslo.utils==3.31.0
packaging==16.8
pbr==3.1.1
pip-review==1.0
pip-tools==1.10.1
positional==1.2.1
psycopg2==2.7.3.2
pycodestyle==2.3.1
pyflakes==1.6.0
pylint==1.7.4
pyparsing==2.2.0
python-dateutil==2.6.1
python-keystoneclient==3.13.0
python-swiftclient==3.4.0
pytz==2017.3
PyYAML==3.12
requests==2.18.4
requests-futures==0.9.7
rfc3986==1.1.0
simplejson==3.12.0
six==1.11.0
sqlparse==0.2.4
stevedore==1.27.1
text-unidecode==1.1
texttable==1.1.1
typing==3.6.2
uritemplate==3.0.0
urllib3==1.21.1
virtualenv==15.1.0
virtualfish==1.0.5
websocket-client==0.44.0
Werkzeug==0.12.2
wrapt==1.10.11
xlsx2csv==0.7.3
